requirement.txt
asgiref==3.5.2
Babel==2.11.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.16
django-environ==0.9.0
django-extensions==3.2.1
django-extra-views==0.13.0
django-haystack==3.2.1
django-oscar==3.1
django-oscar-accounts==3.0.1
django-oscar-api==3.1.2
django-oscar-promotions==1.0.0b1
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
drf-yasg2==1.19.4
factory-boy==3.2.1
Faker==15.1.3
graphviz==0.20.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.12.54
Pillow==9.3.0
purl==1.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.2
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
